Restructure java language server plugin
